requirements.txt
# AI / LLM
groq>=0.9.0

# Data processing
pandas>=2.0.0

# Google Sheets integration
gspread>=6.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
oauth2client>=4.1.3

# Configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Progress display
tqdm>=4.65.0

# Testing
pytest>=7.4.0
